Update readme.
I wasn't really aware of crediting being opt-out and eh, I don't really
care about this bit. Additionally, this is part of a feature branch I
also don't really wish to be credited for, so asking for a credit for a
recolor which is available without such elsewhere would be inconsistent.